reformat code

+++ common-lib/src/main/java/com/NamVu/common/constant/StatusConstant.java
@@ -1,6 +1,6 @@
 package com.NamVu.common.constant;
 
 public class StatusConstant {
-    public static final byte ACTIVE = 1;
-    public static final byte INACTIVE = 0;
+    public static final Integer ACTIVE = 1;
+    public static final Integer INACTIVE = 0;
 }
